+++ src/server.js
@@ -1,3 +1,5 @@
 const app = require("./app");
 
-app.listen(3333);
+app.listen(3333, () => {
+    console.log('Server running on port 3333');
+});
